requirements.txt
aiogram==3.7.0
aiohttp==3.9.5
beautifulsoup4==4.12.3
apscheduler==3.10.4
python-dotenv==1.0.1
lxml==5.2.2
